# docs/swagger.yaml
basePath: /
definitions:
  utility.VerifyTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Reusable Go REST API backend template.
  title: WarSpark API
  version: dev
paths:
...
  /api/v1/clans:
    get:
      parameters:
      - description: 部落名称
        in: query
        name: name
        type: string
      - description: 战争频率
        in: query
        name: warFrequency
        type: string
      - description: 位置ID
        in: query
        name: locationId
        type: integer
      - description: 最少成员数
        in: query
        name: minMembers
        type: integer
      - description: 最多成员数
        in: query
        name: maxMembers
        type: integer
      - description: 最少部落积分
        in: query
        name: minClanPoints
        type: integer
      - description: 最少部落等级
        in: query
        name: minClanLevel
        type: integer
      - description: 分页数量
        in: query
        name: limit
        type: integer
      - description: 游标后
        in: query
        name: after
        type: string
      - description: 游标前
        in: query
        name: before
        type: string
      - description: 标签ID列表
        in: query
        name: labelIds
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 搜索部落
      tags:
      - utility
...
  /api/v1/gold-pass/current:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取当前 Gold Pass
      tags:
      - utility
...
  /api/v1/locations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取位置列表
      tags:
      - ranking
  /api/v1/locations/{id}:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取位置详情
      tags:
      - utility
  /api/v1/locations/{id}/rankings/clans:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落排名
      tags:
      - ranking
  /api/v1/locations/{id}/rankings/clans-builder-base:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落夜世界排名
      tags:
      - ranking
  /api/v1/locations/{id}/rankings/clans-capital:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取部落都城排名
      tags:
      - ranking
  /api/v1/locations/{id}/rankings/players:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家排名
      tags:
      - ranking
  /api/v1/locations/{id}/rankings/players-builder-base:
    get:
      parameters:
      - description: 位置ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家夜世界排名
      tags:
      - ranking
...
  /api/v1/players/{tag}/league-group:
    get:
      parameters:
      - description: 玩家标签
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取玩家联赛分组
      tags:
      - utility
  /api/v1/players/{tag}/verify-token:
    post:
      parameters:
      - description: 玩家标签
        in: path
        name: tag
        required: true
        type: string
      - description: 验证请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utility.VerifyTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 验证玩家令牌
      tags:
      - utility
...
schemes:
- http
swagger: "2.0"